requirements.txt
streamlit==1.28.0
numpy==1.24.0
pandas==2.0.0
opencv-python-headless==4.8.0
Pillow==10.0.0
pyzbar==0.1.9
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
tldextract==5.1.0
qrcode==7.4.0
imutils==0.5.4
python-multipart==0.0.6